# packages/shared/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Viking 4X API
  version: 0.1.0
  description: Server-authoritative API for the Viking 4X client. Contract first; clients are generated from this file. Index of planned routes lives in the roadmap (api-v0.md).
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000
    description: Local development
security:
  - bearer: []
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: {}
    Timer:
      type: object
      required: [id, kind, due_at, state]
      properties:
        id: { type: string, format: uuid }
        kind: { type: string }
        ref_type: { type: [string, "null"] }
        ref_id: { type: [string, "null"], format: uuid }
        due_at: { type: string, format: date-time }
        state: { type: string, enum: [pending, done, cancelled] }
        payload: { type: object }
    Building:
      type: object
      required: [id, kind, slot, level]
      properties:
        id: { type: string, format: uuid }
        kind: { type: string }
        slot: { type: integer }
        level: { type: integer }
    MapHall:
      type: object
      required: [x, y, hall_id, name, level, shielded]
      properties:
        x: { type: integer }
        y: { type: integer }
        hall_id: { type: string, format: uuid }
        player_id: { type: string, format: uuid }
        name: { type: string }
        level: { type: integer }
        shielded: { type: boolean }
    TerrainChunk:
      description: 64x64 terrain tiles, one byte each, base64. Tile byte:0 land, 1 coast, 2 sea, 3 mountain. Immutable for the life of the kingdom.
      type: object
      required: [cx, cy, x0, y0, w, h, tiles]
      properties:
        cx: { type: integer }
        cy: { type: integer }
        x0: { type: integer }
        y0: { type: integer }
        w: { type: integer }
        h: { type: integer }
        tiles: { type: string, contentEncoding: base64 }
paths:
  /health:
    get:
      operationId: getHealth
      summary: Health check
      security: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [ok, service]
                properties:
                  ok: { type: boolean }
                  service: { type: string }
  /v1/auth/guest:
    post:
      operationId: authGuest
      summary: Create or resume a guest account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [device_id]
              properties:
                device_id: { type: string, minLength: 8 }
      responses:
        "200":
          description: JWT and account
          content:
            application/json:
              schema:
                type: object
                required: [jwt, account, server_now]
                properties:
                  jwt: { type: string }
                  account: { type: object }
                  player: { type: [object, "null"] }
                  server_now: { type: string, format: date-time }
        "400":
          description: Bad device id
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /v1/players:
    post:
      operationId: signUp
      summary: Sign up into the newest open kingdom; places the hall
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 2, maxLength: 20 }
      responses:
        "200":
          description: Player, hall position, new JWT
          content:
            application/json:
              schema:
                type: object
                required: [player, jwt, server_now]
                properties:
                  player: { type: object }
                  jwt: { type: string }
                  server_now: { type: string, format: date-time }
        "409":
          description: Already in kingdom, or kingdom full
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /v1/hall:
    get:
      operationId: getHall
      summary: My hall, buildings, and pending timers
      responses:
        "200":
          description: Hall
          content:
            application/json:
              schema:
                type: object
                required: [hall, buildings, timers, server_now]
                properties:
                  hall: { type: object }
                  buildings: { type: array, items: { $ref: "#/components/schemas/Building" } }
                  timers: { type: array, items: { $ref: "#/components/schemas/Timer" } }
                  server_now: { type: string, format: date-time }
        "401":
          description: Unauthenticated
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /v1/buildings/{id}/upgrade:
    post:
      operationId: upgradeBuilding
      summary: Start an upgrade (Longhouse-gated, two builders, costs resources)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Timer started
          content:
            application/json:
              schema:
                type: object
                required: [timer, server_now]
                properties:
                  timer: { $ref: "#/components/schemas/Timer" }
                  server_now: { type: string, format: date-time }
        "409":
          description: No free builder, or already upgrading
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "422":
          description: Longhouse gate, max level, or insufficient resources
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /v1/map/chunk:
    get:
      operationId: getTerrainChunk
      summary: One 64x64 terrain chunk; immutable, cache forever
      parameters:
        - name: cx
          in: query
          required: true
          schema: { type: integer, minimum: 0 }
        - name: cy
          in: query
          required: true
          schema: { type: integer, minimum: 0 }
      responses:
        "200":
          description: Terrain chunk
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TerrainChunk" }
        "400":
          description: Chunk outside the kingdom
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /v1/map/viewport:
    get:
      operationId: getViewport
      summary: Occupants in a rectangle (max 64x64 tiles)
      parameters:
        - { name: x0, in: query, required: true, schema: { type: integer } }
        - { name: y0, in: query, required: true, schema: { type: integer } }
        - { name: x1, in: query, required: true, schema: { type: integer } }
        - { name: y1, in: query, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Occupants in the clamped rectangle
          content:
            application/json:
              schema:
                type: object
                required: [bounds, halls, server_now]
                properties:
                  bounds:
                    type: object
                    required: [x0, y0, x1, y1]
                    properties:
                      x0: { type: integer }
                      y0: { type: integer }
                      x1: { type: integer }
                      y1: { type: integer }
                  halls: { type: array, items: { $ref: "#/components/schemas/MapHall" } }
                  nodes: { type: array, items: { type: object } }
                  camps: { type: array, items: { type: object } }
                  holds: { type: array, items: { type: object } }
                  banners: { type: array, items: { type: object } }
                  marches: { type: array, items: { type: object } }
                  server_now: { type: string, format: date-time }
        "400":
          description: Viewport too large
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /v1/map/overview.png:
    get:
      operationId: getKingdomOverview
      summary: The whole kingdom as one indexed PNG, one pixel per tile; immutable, cache forever
      responses:
        "200":
          description: Kingdom image
          content:
            image/png:
              schema: { type: string, format: binary }
        "401":
          description: Unauthenticated
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /v1/map/overview:
    get:
      operationId: getKingdomOverviewMarkers
      summary: The live layer drawn over the kingdom image — halls now, holds and territory from P4
      responses:
        "200":
          description: Markers
          content:
            application/json:
              schema:
                type: object
                required: [size, halls, server_now]
                properties:
                  size: { type: integer, description: Kingdom width in tiles; the image is this many pixels square }
                  halls:
                    type: array
                    items:
                      type: object
                      required: [hall_id, x, y, name, own]
                      properties:
                        hall_id: { type: string, format: uuid }
                        x: { type: integer }
                        y: { type: integer }
                        name: { type: string }
                        tribe_id: { type: [string, "null"], format: uuid }
                        own: { type: boolean }
                  holds: { type: array, items: { type: object } }
                  territory: { type: array, items: { type: object } }
                  pins: { type: array, items: { type: object } }
                  server_now: { type: string, format: date-time }
        "401":
          description: Unauthenticated
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }